requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
cssselect==1.1.0
dnspython==2.1.0
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
lxml==4.6.4
packaging==21.3
parse==1.19.0
pluggy==1.0.0
py==1.11.0
pyee==8.2.2
pymongo==4.0
pyparsing==3.0.6
pyppeteer==0.2.6
pyquery==1.4.3
pytest==6.2.5
pytest-timeout==2.0.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
zipp==3.6.0
